Refactor auth exception and auth method
