+++ src/zepben/auth/common/auth_exception.py
@@ -1,0 +1,11 @@
+#  Copyright $year Zeppelin Bend Pty Ltd
+#
+#  This Source Code Form is subject to the terms of the Mozilla Public
+#  License, v. 2.0. If a copy of the MPL was not distributed with this
+#  file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+__all__ = ["AuthException"]
+
+
+class AuthException(Exception):
+    pass
